The proxy polls the main process and kills itself if it doesnt hear back
